requirements.txt
ccxt
pandas-ta
websocket-client